CSS-in-JS: Success modal
